LayoutTests/imported/w3c/web-platform-tests/html/browsers/browsing-the-web/history-traversal/same-url-expected.txt


FAIL Test same-URL navigation and its effects on history assert_equals: expected 1 but got 100
